rename mix closure
